Update remaining distance references to double

+++ runkit/fusedapp/src/main/java/com/joshualorett/fusedapp/DistanceFormatter.kt
@@ -4,7 +4,7 @@ package com.joshualorett.fusedapp
  * Formatted distance text.
  * Created by Joshua on 10/18/2020.
  */
-fun formatDistance(distance: Float): String {
+fun formatDistance(distance: Double): String {
     return if (distance < 1000) {
         val roundedDistance = "%.2f".format(distance)
         "$roundedDistance m"
